Moved any prompts out of OnyxClient api class, and testing typer CLI

+++ onyx/config.py
@@ -1,9 +1,9 @@
 import os
-import stat  # TODO: Create tokens file with correct file permissions
 import json
-from typing import Any, List, Dict, Tuple, Optional, Union
+from typing import Optional, Union
 
 # TODO: Appropriate error messages for stuff such as domain not existing (i.e. when there is no config file)
+# TODO: Ok as first step. refactor to do what Andy said and pull out file management
 
 
 class OnyxConfig:
@@ -14,13 +14,11 @@ class OnyxConfig:
         self,
         config_path: Optional[str] = None,
         domain: Optional[str] = None,
+        token: Optional[str] = None,
         username: Optional[str] = None,
         password: Optional[str] = None,
-        token: Optional[str] = None,
     ) -> None:
         self.domain = domain
-        self.username = username
-        self.password = password
         self.token = token
 
         if config_path:
@@ -33,13 +31,18 @@ class OnyxConfig:
         if os.path.isfile(self.config_path):
             with open(self.config_path) as config_file:
                 config = json.load(config_file)
-                if not self.domain:
-                    self.domain = config["domain"]
 
-                if not self.token:
-                    self.token = config["token"]
+                if not self.domain and config.get("domain"):
+                    self.domain = str(config["domain"])
+
+                if not self.token and config.get("token"):
+                    self.token = str(config["token"])
         else:
-            pass  # TODO: Check for domain, if not raise error cause this is needed
+            if not self.domain:
+                raise Exception("Could not find domain name.")
+
+        self.username = username
+        self.password = password
 
     def write_token(self, token: Union[str, None]) -> None:
         """Update the tokens file for `username`.
@@ -50,12 +53,14 @@ class OnyxConfig:
             The token being written to their tokens file.
         """
 
-        with open(self.config_path, "w") as config_file:
-            json.dump(
-                {
+        if os.path.isfile(self.config_path):
+            with open(self.config_path, "w") as config_file:
+                config = {
                     "domain": self.domain,
                     "token": token,
-                },
-                config_file,
-                indent=4,
-            )
+                }
+                json.dump(
+                    config,
+                    config_file,
+                    indent=4,
+                )
